// src/client/models/OfficeLinkTypeEnum.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * An enumeration.
 */
export enum OfficeLinkTypeEnum {
    GOOGLE_MAPS = 'GOOGLE_MAPS',
    YANDEX_MAPS = 'YANDEX_MAPS',
    DUBLE_GIS_MAPS = 'DUBLE_GIS_MAPS',
    FACEBOOK = 'FACEBOOK',
    VK = 'VK',
    INSTAGRAM = 'INSTAGRAM',
    WHATSAPP = 'WHATSAPP',
    TELEGRAM = 'TELEGRAM',
    VIBER = 'VIBER',
    TIKTOK = 'TIKTOK',
    OTHER = 'OTHER',
}